+++ src/components/App.jsx
@@ -439,7 +439,7 @@ class App extends Component {
 						fetch("/SemanticViewer");
 					}}
 				>
-					Semantics Viewer
+					Dataset Type Viewer
 				</MenuItem>
 			</Menu>
 		);